+++ src/launcher/util/registerIpcHandler.ts
@@ -7,6 +7,7 @@
 import { ErrorDialogActions } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
 import * as appInstallProgress from '../../ipc/appInstallProgress';
+import * as jlink from '../../ipc/installJLink';
 import * as launcherUpdateProgress from '../../ipc/launcherUpdateProgress';
 import * as proxyLogin from '../../ipc/proxyLogin';
 import * as showErrorDialog from '../../ipc/showErrorDialog';
@@ -14,6 +15,7 @@ import {
     initialiseAppInstallProgress,
     updateAppInstallProgress,
 } from '../features/apps/appsSlice';
+import { updateProgress as updateJLinkProgress } from '../features/jlinkUpdate/jlinkUpdateSlice';
 import {
     reset,
     startDownload,
@@ -46,5 +48,8 @@ export default (dispatch: AppDispatch) => {
     });
     launcherUpdateProgress.forMain.registerUpdateFinished(() => {
         dispatch(reset());
+    });
+    jlink.forMain.registerUpdateJLinkProgress(update => {
+        dispatch(updateJLinkProgress(update));
     });
 };