+++ src/main/bundledJlink.d.ts
@@ -4,6 +4,6 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-declare const bundledJlinkVersion: string;
+const bundledJlinkVersion: string;
 
 export default bundledJlinkVersion;