+++ src/main/configureElectronApp.ts
@@ -7,7 +7,6 @@
 import { app, dialog, Menu } from 'electron';
 import fs from 'fs';
 import os from 'os';
-import path from 'path';
 
 import {
     getBundledAppInstalled,
@@ -24,9 +23,10 @@ import {
     getAppsExternalDir,
     getAppsLocalDir,
     getAppsRootDir,
-    getBundledResourcesDir,
+    getBundledResourcePath,
     getNodeModulesDir,
-    getUserDataDir,
+    getUnpackedBundledResourcePath,
+    getUserDataPath,
 } from './config';
 import describeError from './describeError';
 import loadDevtools from './devtools';
@@ -87,23 +87,19 @@ const fatalError = (error: unknown) => {
 const copyNrfutil = () => {
     const binName = `nrfutil${process.platform === 'win32' ? '.exe' : ''}`;
 
-    const nrfutilBundled = path.join(getBundledResourcesDir(), binName);
-    const nrfutilInAppPath = path.join(getUserDataDir(), binName);
+    const nrfutilBundled = getBundledResourcePath(binName);
+    const nrfutilInAppPath = getUserDataPath(binName);
 
     fs.copyFileSync(nrfutilBundled, nrfutilInAppPath);
 };
 
 const copyNrfutilSandboxes = async () => {
-    const nrfutilBundledSandboxes = path
-        .join(getBundledResourcesDir(), 'nrfutil-sandboxes')
-        .replace('app.asar', 'app.asar.unpacked');
+    const nrfutilBundledSandboxes =
+        getUnpackedBundledResourcePath('nrfutil-sandboxes');
 
     if (!fs.existsSync(nrfutilBundledSandboxes)) return;
 
-    const nrfutilBundledSandboxesDest = path.join(
-        getUserDataDir(),
-        'nrfutil-sandboxes'
-    );
+    const nrfutilBundledSandboxesDest = getUserDataPath('nrfutil-sandboxes');
 
     fs.mkdirSync(nrfutilBundledSandboxesDest, { recursive: true });